+++ packages/api/src/Domain/Request/Workspace/WorkspaceInvitationRequestParams.ts
@@ -1,0 +1,8 @@
+import { Uuid, WorkspaceAccessLevel } from '@standardnotes/common'
+
+export type WorkspaceInvitationRequestParams = {
+  workspaceUuid: Uuid
+  inviteeEmail: string
+  accessLevel: WorkspaceAccessLevel
+  [additionalParam: string]: unknown
+}