+++ packages/api/src/Domain/Server/Workspace/Paths.ts
@@ -1,5 +1,11 @@
+import { Uuid } from '@standardnotes/common'
+
 const WorkspacePaths = {
   createWorkspace: '/v1/workspaces',
+  listWorkspaces: '/v1/workspaces',
+  listWorkspaceUsers: (uuid: Uuid) => `/v1/workspaces/${uuid}/users`,
+  inviteToWorkspace: (uuid: Uuid) => `/v1/workspaces/${uuid}/invites`,
+  acceptInvite: (uuid: Uuid) => `/v1/invites/${uuid}/accept`,
 }
 
 export const Paths = {