+++ packages/services/src/Domain/Workspace/WorkspaceManager.ts
@@ -1,4 +1,7 @@
 import { WorkspaceApiServiceInterface } from '@standardnotes/api'
+import { Uuid, WorkspaceAccessLevel, WorkspaceType } from '@standardnotes/common'
+import { Workspace, WorkspaceUser } from '@standardnotes/models'
+
 import { InternalEventBusInterface } from '../Internal/InternalEventBusInterface'
 import { AbstractService } from '../Service/AbstractService'
 import { WorkspaceClientInterface } from './WorkspaceClientInterface'
@@ -11,10 +14,76 @@ export class WorkspaceManager extends AbstractService implements WorkspaceClient
     super(internalEventBus)
   }
 
-  async createWorkspace(dto: {
-    encryptedWorkspaceKey: string
-    encryptedPrivateKey: string
+  async listWorkspaceUsers(dto: { workspaceUuid: string }): Promise<{ users: WorkspaceUser[] }> {
+    try {
+      const result = await this.workspaceApiService.listWorkspaceUsers(dto)
+
+      if (result.data.error !== undefined) {
+        return { users: [] }
+      }
+
+      return result.data
+    } catch (error) {
+      return { users: [] }
+    }
+  }
+
+  async listWorkspaces(): Promise<{ ownedWorkspaces: Workspace[]; joinedWorkspaces: Workspace[] }> {
+    try {
+      const result = await this.workspaceApiService.listWorkspaces()
+
+      if (result.data.error !== undefined) {
+        return { ownedWorkspaces: [], joinedWorkspaces: [] }
+      }
+
+      return result.data
+    } catch (error) {
+      return { ownedWorkspaces: [], joinedWorkspaces: [] }
+    }
+  }
+
+  async acceptInvite(dto: {
+    inviteUuid: string
+    userUuid: string
     publicKey: string
+    encryptedPrivateKey: string
+  }): Promise<{ success: boolean }> {
+    try {
+      const result = await this.workspaceApiService.acceptInvite(dto)
+
+      if (result.data.error !== undefined) {
+        return { success: false }
+      }
+
+      return result.data
+    } catch (error) {
+      return { success: false }
+    }
+  }
+
+  async inviteToWorkspace(dto: {
+    inviteeEmail: string
+    workspaceUuid: Uuid
+    accessLevel: WorkspaceAccessLevel
+  }): Promise<{ uuid: string } | null> {
+    try {
+      const result = await this.workspaceApiService.inviteToWorkspace(dto)
+
+      if (result.data.error !== undefined) {
+        return null
+      }
+
+      return result.data
+    } catch (error) {
+      return null
+    }
+  }
+
+  async createWorkspace(dto: {
+    workspaceType: WorkspaceType
+    encryptedWorkspaceKey?: string
+    encryptedPrivateKey?: string
+    publicKey?: string
     workspaceName?: string
   }): Promise<{ uuid: string } | null> {
     try {