+++ packages/files/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.15](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.14...@standardnotes/files@1.10.15) (2022-10-11)
+
+**Note:** Version bump only for package @standardnotes/files
+
+## [1.10.14](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.13...@standardnotes/files@1.10.14) (2022-10-10)
+
+**Note:** Version bump only for package @standardnotes/files
+
+## [1.10.13](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.12...@standardnotes/files@1.10.13) (2022-10-10)
+
+### Bug Fixes
+
+* **api:** workspace creation arguments ([a275a45](https://github.com/standardnotes/app/commit/a275a45753abc4a29a89de5fe784260165297bbe))
+
 ## [1.10.12](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.11...@standardnotes/files@1.10.12) (2022-10-06)
 
 **Note:** Version bump only for package @standardnotes/files