+++ packages/services/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.29.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.28.0...@standardnotes/services@1.29.0) (2022-10-11)
+
+### Features
+
+* **api:** add access level to invitation requests ([49b3d27](https://github.com/standardnotes/app/commit/49b3d2714fbe4ac9f9242097b8df6b4e3d604513))
+
+# [1.28.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.27.1...@standardnotes/services@1.28.0) (2022-10-11)
+
+### Features
+
+* **api:** add accepting invites to workspace ([15e2c82](https://github.com/standardnotes/app/commit/15e2c82e654ce3f983cd42b01f14d18435b3e939))
+* **api:** add listing workspaces ([8376381](https://github.com/standardnotes/app/commit/837638198112df24c30e53803a7ddf5f841975b0))
+
+## [1.27.1](https://github.com/standardnotes/app/compare/@standardnotes/services@1.27.0...@standardnotes/services@1.27.1) (2022-10-10)
+
+### Bug Fixes
+
+* dark mode not working in editors ([#1773](https://github.com/standardnotes/app/issues/1773)) ([27d2c95](https://github.com/standardnotes/app/commit/27d2c95b5b82645bf6c0372979f05236c83a10f6))
+
+# [1.27.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.26.0...@standardnotes/services@1.27.0) (2022-10-10)
+
+### Bug Fixes
+
+* **api:** workspace creation arguments ([a275a45](https://github.com/standardnotes/app/commit/a275a45753abc4a29a89de5fe784260165297bbe))
+* linter issues ([e2e7605](https://github.com/standardnotes/app/commit/e2e76054142e8d67cac69280e2a3efcb15cf67f1))
+
+### Features
+
+* **api:** add inviting to workspace ([158ca6a](https://github.com/standardnotes/app/commit/158ca6ac6a428e082c8f984791f80fcaadcb3e34))
+
 # [1.26.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.25.0...@standardnotes/services@1.26.0) (2022-10-07)
 
 ### Features